move route planning into model
Relocate template dependency, route graph, and commitment planning out of
the emitter while preserving model construction and generated output.
